+++ apps/web/app/(use-page-wrapper)/(main-nav)/teams/server-page.tsx
@@ -5,7 +5,7 @@ import { unstable_cache } from "next/cache";
 import { checkAdminOrOwner } from "@calcom/features/auth/lib/checkAdminOrOwner";
 import { TeamsListing } from "@calcom/features/ee/teams/components/TeamsListing";
 import { TeamRepository } from "@calcom/lib/server/repository/team";
-import { TeamService } from "@calcom/lib/server/service/team";
+import { TeamService } from "@calcom/lib/server/service/teamService";
 import prisma from "@calcom/prisma";
 
 import { TRPCError } from "@trpc/server";
