+++ packages/trpc/server/routers/viewer/teams/delete.handler.ts
@@ -1,5 +1,5 @@
 import { isTeamOwner } from "@calcom/lib/server/queries/teams";
-import { TeamService } from "@calcom/lib/server/service/team";
+import { TeamService } from "@calcom/lib/server/service/teamService";
 
 import { TRPCError } from "@trpc/server";
